refactor: MonthNav 적용
